refactor(client): extract InlineVideoPlayer + FullscreenVideoPlayer to dedicated file
media_content.dart had grown to 1467 LOC because both inline-video
state machines lived inside it alongside the image / file / audio /
GIF rendering branches. They're self-contained — neither references
private state of MediaContent — so move them out cleanly.

- New file lib/src/widgets/message/video_player.dart holds
  InlineVideoPlayer (+ _PlayOverlay) and FullscreenVideoPlayer
  (+ _FullscreenVideoPlayerState). 623 LOC.
- media_content.dart drops to 854 LOC (-42%) and loses its dependency
  on media_kit / media_kit_video imports — those move to the new file
  with the consumers.
- Test imports updated: media_content_video_test now imports
  video_player.dart explicitly for the InlineVideoPlayer symbol.

Pure mechanical move; no logic changes; all 1332 tests pass.
